Moved mergeCaseLinks() method into CaseLinkService class
